models/text-to-image/requirements-amd.txt
transformers
diffusers
accelerate
scipy
safetensors
datasets
pillow
peft
argcomplete
requests
protobuf
sentencepiece
